// web/src/gen/model/analyzedCustomNodeDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfPinDefinition } from "./iReadOnlyListOfPinDefinition";
import type { IReadOnlyListOfPinDiffItem } from "./iReadOnlyListOfPinDiffItem";
import type { ScriptImpactReportDto } from "./scriptImpactReportDto";

export interface AnalyzedCustomNodeDto {
  fileName: string;
  key: string;
  suggestedName: string;
  suggestedLabel: string;
  executor: string;
  /** @nullable */
  description: string | null;
  contentHash: string;
  isOverride: boolean;
  /** @nullable */
  existingNodeId: string | null;
  inputs: IReadOnlyListOfPinDefinition;
  outputs: IReadOnlyListOfPinDefinition;
  inputPinDiffs: IReadOnlyListOfPinDiffItem;
  outputPinDiffs: IReadOnlyListOfPinDiffItem;
  impactReport: ScriptImpactReportDto;
}